% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geo.R
\name{get_geo_prefix}
\alias{get_geo_prefix}
\title{Generate GEO Prefix}
\usage{
get_geo_prefix(geo_string)
}
\arguments{
\item{geo_string}{A character string, typically a GSE accession number.}
}
\value{
A character string representing the GEO prefix.
}
\description{
Creates a prefix string for constructing GEO FTP paths.
}
\examples{
# get_geo_prefix("GSE12345") # GSE12nnn
# get_geo_prefix("GSE123")   # GSEnnn
}
\keyword{internal}
